factory reset POC
